+++ src/net/lodoma/lime/world/server/ServersideWorld.java
@@ -35,8 +35,8 @@ public class ServersideWorld implements TileGrid, EntityWorld
     private boolean paletteLock;
     
     private PhysicsWorld physicsWorld;
-    private Map<Long, Entity> entities;
-    private List<Long> entitiesToCreate;
+    private Map<Integer, Entity> entities;
+    private List<Integer> entitiesToCreate;
     
     public ServersideWorld(Server server)
     {
@@ -44,8 +44,8 @@ public class ServersideWorld implements TileGrid, EntityWorld
         this.palette = new HashMap<Short, Material>();
         
         physicsWorld = new PhysicsWorld();
-        entities = new HashMap<Long, Entity>();
-        entitiesToCreate = new ArrayList<Long>();
+        entities = new HashMap<Integer, Entity>();
+        entitiesToCreate = new ArrayList<Integer>();
     }
     
     @Override
@@ -189,7 +189,7 @@ public class ServersideWorld implements TileGrid, EntityWorld
                 chunks[y * chunkAX + x].lockState();
     }
     
-    public void createEntity(long id)
+    public void createEntity(int id)
     {
         entitiesToCreate.add(id);
     }
@@ -202,26 +202,26 @@ public class ServersideWorld implements TileGrid, EntityWorld
     }
 
     @Override
-    public Entity getEntity(long id)
+    public Entity getEntity(int id)
     {
         return entities.get(id);
     }
     
-    public Set<Long> getEntityIDSet()
+    public Set<Integer> getEntityIDSet()
     {
-        return new HashSet<Long>(entities.keySet());
+        return new HashSet<Integer>(entities.keySet());
     }
 
     @Override
-    public void removeEntity(long id)
+    public void removeEntity(int id)
     {
         entities.remove(id);
     }
     
     public void update(double timeDelta)
     {
-        List<Long> createdEntities = new ArrayList<Long>();
-        for(Long entityID : entitiesToCreate)
+        List<Integer> createdEntities = new ArrayList<Integer>();
+        for(int entityID : entitiesToCreate)
         {
             entities.get(entityID).create(physicsWorld);
             createdEntities.add(entityID);
